V3.6.1 oh network (#13052)
* openharmony network support

* Revert "Oh network adapter (#13013)"

This reverts commit 6a3849f137b9bcf183e138edf9139f9cf071a4e0.

+++ native/cocos/network/HttpRequest.h
@@ -242,8 +242,8 @@ public:
         _timeoutInSeconds = timeoutInSeconds;
     }
 
-    inline float getTimeout() const {
-        return _timeoutInSeconds;
+    inline long getTimeout() const {
+        return static_cast<long>(_timeoutInSeconds);
     }
 
 protected:
